fixes: fixes after refactor
