+++ src/app/features/private/community/components/group-list/group-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, input } from '@angular/core';
-import { GroupCardComponent } from '../../../shared/components/group-card/group-card.component';
+import { GroupCardComponent } from '../../../shared/group-card/group-card.component';
 import { CommonModule } from '@angular/common';
 import { Group } from '../../../../shared/models/group/group.model';
 import { NzGridModule } from 'ng-zorro-antd/grid';
