+++ src/app/features/private/dashboard/components/dashboard-activities/dashboard-activities.component.ts
@@ -1,6 +1,6 @@
 import { Component, input, output } from '@angular/core';
 import { ActivityItemDetails } from '../../../../shared/models/task-models/activity.model';
-import { ActivityItemComponent } from '../../../shared/components/task-item/activity-item.component';
+import { ActivityItemComponent } from '../../../shared/task-item/activity-item.component';
 import { NzFlexDirective } from 'ng-zorro-antd/flex';
 import { NzListComponent, NzListEmptyComponent } from 'ng-zorro-antd/list';
 import { NzButtonModule } from 'ng-zorro-antd/button';
