+++ src/app/features/private/group-members/group-members-page/group-members-page.component.ts
@@ -18,7 +18,7 @@ import { NzButtonModule } from 'ng-zorro-antd/button';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 import { NotificationService } from '../../../shared/services/notification-service/notification.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { BackButtonComponent } from '../../shared/components/back-button/back-button-component';
+import { BackButtonComponent } from '../../shared/back-button/back-button-component';
 
 @Component({
   selector: 'app-group-members',
