+++ src/app/features/private/group-preview/components/group-shop-view/group-shop/group-shop.component.ts
@@ -11,7 +11,7 @@ import {
 import { Group } from '../../../../../shared/models/group/group.model';
 import { GroupItemModel } from '../../../../../shared/models/group/group-item.model';
 import { GroupShopApiService } from '../../../../../shared/services/group-shop-api/group-shop-api.service';
-import { PaginationMoreComponent } from '../../../../shared/components/pagination-more/pagination-more.component';
+import { PaginationMoreComponent } from '../../../../shared/pagination-more/pagination-more.component';
 import { GroupPreviewMode } from '../../../../../shared/models/group/group-preview-mode';
 import { NzSpinComponent } from 'ng-zorro-antd/spin';
 import { GroupItemComponent } from '../group-item/group-item.component';
