+++ src/app/features/private/group-ranking/group-ranking-page/group-ranking-page.component.ts
@@ -11,7 +11,7 @@ import { GroupMember } from '../../../shared/models/group/group-member.model';
 import { FullRankingComponent } from '../components/full-ranking/full-ranking.component';
 import { NotificationService } from '../../../shared/services/notification-service/notification.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { BackButtonComponent } from '../../shared/components/back-button/back-button-component';
+import { BackButtonComponent } from '../../shared/back-button/back-button-component';
 
 @Component({
   selector: 'app-group-ranking',
