+++ src/app/features/private/group-requests/group-requests-page/group-requests-page.component.ts
@@ -10,7 +10,7 @@ import { NzModalModule } from 'ng-zorro-antd/modal';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 import { NotificationService } from '../../../shared/services/notification-service/notification.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { BackButtonComponent } from '../../shared/components/back-button/back-button-component';
+import { BackButtonComponent } from '../../shared/back-button/back-button-component';
 
 @Component({
   selector: 'app-group-requests',
