+++ src/app/features/private/shared/group-card/group-card.component.css
