+++ src/app/features/private/shared/group-type-tag/group-type-tag.component.ts
